flashfiles iterative refinement
At build time, we revert back to using BOARD_FLASHFILES, which is
a list of colon-separated filename pairs. The first item in the
pair is the actual path to the file in the build system and can
be anything. The second item is the name of the file inside
the TFP's RADIO/flashfiles.zip, and must correspond to the "target"
field in the JSON configuration.

Change-Id: I3c19a4c951b641e4037c09f117f4f4d4bb5ac6da
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Reviewed-on: https://android.intel.com/274438