add prebuilt cache and userdata partition
Cache is fixed at 256M in GMIN-L. Provide a prebuilt one so that
new devices don't have to format it at first boot.

Userdata partition should also be set, but since the size of this
is very board-specific, it has to be set in individual BoardConfig.mk.
Emit a warning if this isn't done.

By pre-populating these partitions, the device doesn't have to
reboot itself after being provisioned in the factory.

setup_fs service removed; fs_mgr initiates a reboot into Recovery
to format if it detects if userdata is zeroed out,
which happens after unlocking/locking device state in Fastboot.

Change-Id: Idf870d76e31da9d865ad0bc7daf8d11c6ed081a2
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Reviewed-on: https://android.intel.com/277597