Start dhcpcd on post-fs-data
When started on post-fs event, the dhcpcd service gets killed by vold
when trying to unmount the data partition before encrypting it. This happens
because dhcpcd still has open handles on the data partition
(/data/misc/dhcp/dhcpcd-eth0.pid).
To bypass this problem start dhcpcd on post-fs-data event to make sure
that the data partition is mounted and encrypted.

Change-Id: I8fb97fe2e97ea094ce72e65765871c7367a56f2d
Depends-Change-Id: I5684ea612332ae810cf754ddd8afa7de0e2d2c2f
Tracked-By: GMINL-3969
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-3969
Signed-off-by: Petre Pircalabu <petre.ovidiu.pircalabu@intel.com>
Reviewed-on: https://android.intel.com/302643
Tested-by: cactus <cactus@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Dolca, Robert <robert.dolca@intel.com>
Reviewed-on: https://android.intel.com/335479
Reviewed-by: Serban, Mihai <mihai.serban@intel.com>
Tested-by: Dolca, Robert <robert.dolca@intel.com>
Reviewed-by: Dalleau, Frederic <frederic.dalleau@intel.com>
Tested-by: Dalleau, Frederic <frederic.dalleau@intel.com>