graphics/ufo_gen7: init.rc.1: Setting autosuspend_delay_ms to 500ms for VLV
RPM autosuspend delay was set to 500ms in MCG codebase. With GMIN, default
value of 10000ms is very high and we keep display wells ON for this time.
This will drain power. To avoid this autosuspend_delay is reduced to 500ms.
By this time all timers are cancelled and if any shrinker calls runtime resume
is triggered.

v4: Changed sysfs name autosuspend_delay to autosuspend_delay_ms. (Sagar)

Issue: GMINL-2847
Change-Id: I076e795e6763eea9fe45b5a6cc77d678b2583db9
Signed-off-by: Sagar Kamble <sagar.a.kamble@intel.com>
Reviewed-on: https://android.intel.com/305877
Reviewed-by: Lee, Chang-joon <chang-joon.lee@intel.com>
Tested-by: Lee, Chang-joon <chang-joon.lee@intel.com>