Moved config partition from telephony to config-partition mixin
Config partition is used to setup some configuration to the system.
Since this partition is used by telephony but also by other FT, it
does not make sense to have it in telephony mixin. It should be
stored in a new mixin (config-partition) where it can be enabled if
the built target needs it.

Change-Id: Ic575be4ce6f20411dad1c6640bdb9b369a10080d
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-5225
Signed-off-by: Cyril Marpaud <cyrilx.marpaud@intel.com>
Reviewed-on: https://android.intel.com/321165
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Compostella, Jeremy <jeremy.compostella@intel.com>
Reviewed-by: Bessiere, Sebastien <sebastien.bessiere@intel.com>
Reviewed-by: Boie, Andrew P <andrew.p.boie@intel.com>
Tested-by: Boie, Andrew P <andrew.p.boie@intel.com>