art-config: adjust policy
The current policy of excluding dexpreopt in userdebug builds creates
a large difference in required disk space and first-time boot performance
between user and userdebug builds. This has led to some nasty surprises
especially with respect to system partition disk usage.

This is highly undesirable. The differences between user and userdebug are
intended to be very minimal; the documentation on source.android.com for this
describes the build types as follows:

user - limited access; suited for production
userdebug - like "user" but with root access and debuggability; preferred
            for debugging
eng - development configuration with additional debugging tools

https://source.android.com/source/building-running.html

This patch changes the policy to omit dexpreopt in 'eng' builds and leave it
in for user/userdebug.

Issue: GMINL-5629
Change-Id: Iba763baa3e39cdb44fa2ea634ec475f7b07aaff1
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Reviewed-on: https://android.intel.com/333707
Reviewed-by: Fagerstedt, Axel <axel.fagerstedt@intel.com>
Tested-by: Fagerstedt, Axel <axel.fagerstedt@intel.com>