change semantics of RSCI boot flag
This flag was only intended for bring-up phase on devices
that report bad data in RSCI.

For devices that don't have RSCI, this situation is already
detected and handled at runtime.

Instead of having to manually turn this on for every board,
just disable it for boards that don't work properly.

Change-Id: I80b2f5b77901cbc2b6ad511ae710f5fd0d2329e0
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Reviewed-on: https://android.intel.com/364512